MAD_HOT_MAIN/backend/requirements.txt
fastapi
uvicorn[standard]
joblib
numpy
pandas
pydantic
email-validator
pymongo
river
scapy
scikit_learn
uvicorn
python-dotenv
python-multipart
kafka-python
bcrypt
python-jose[cryptography]
